Refactor shell to enable streaming of output over http
The shell now uses the hidden iframe technique to stream the
command output. On the server side it's implemented as a WSGI
middleware that intercepts the relevant calls, executes the
command as a subprocess using fabric and prints the results within
<script> tags that append the new lines in the output div.

It appears to work on every browser but requires further testing
especially when the ssh_user has to be something other than root
(e.g. ec2 AMI's)