The (new) core alerting service
This is the core logic of the new alerting service.

The main idea behind the new alerting service is the usage of
plugins in order to allow instances of mist.api.rules.models.Rules
to be evaluated against various backend storages.

This new service can be potentially used for alerting on both
monitoring data, logs, etc.

Only Graphite is currently supported as a backend