Avoid unintentional dereferencing caused by conditions.models.ConditionalClassMixin
Use the Owner's plain ID, instead of an mist.api.users.models.Owner instance,
when fetching resource to avoid unnecessary dereferencing of fields.

This will currently impact the RuleScheduler, as the Rule model already
stores the Owner's ID as a plain string. We should consider applying the
same logic in the UserScheduler.