Increase the UPDATE_INTERVAL setting of all celery schedulers
This commit affects the

  * UserScheduler
  * RuleScheduler
  * PollingScheduler

The UPDATE_INTERVAL setting denotes how often schedules should be
synced with the backend database. By default, the UPDATE_INTERVAL
is set to 5 seconds, meaning every 5 seconds all documents are saved
to and re-loaded from mongodb.

Having such a frequency sync interval has proven to induce a CPU
bottleneck as the number of documents increases.