Use clouds' UUID instead of byte string in celery schedules' name
of CloudPollingSchedule

Later on, this will allow us to apply the FIXME in order to avoid
dereferencing the `cloud` field everytime we call save(). Since the
schedule's name we only ever be set once, we need to make sure that
it is derived from a unique identifier in order to avoid uniqueness
constrains.