Mark machines, whose cloud has been marked as deleted, as missing
This is necessary in order for `MachinePollingSchedule`s to stop
being scheduled for machines, whose cloud has been deleted. This
change also helps us simplify the `mist.api.tasks:gc_schedulers`
task, since we do not have to explicitly check for deleted clouds
in case of `MachinePollingSchedule`s.

We also ensure that a `ListMachinesPollingSchedule` is in place in
case a cloud is re-enabled, since schedules of disabled clouds are
GCed.

A simple migration script is included to set the `missing_since`
field of machines to the clouds' deletion date, if not already set.