Update the `NonDeleteSchedule` class passed to the UserScheduler
The intial subclass has been re-written into a simple wrapper class. This
allows to trick the scheduler by providing an interface similar to that of
a mongoengine Document subclass in order to prevent schedules marked as deleted
from being loaded.

Similarly, we could have used a custom QuerySet manager to achieve this, which
was the initial approach to solve the problem. However, subclasses of mongoengine
models, which are not a direct subclass of the main `Document` class, do not
fetch the documents of the corresponding superclass. In that case, we'd have to
override mongoengine's QuerySet class in a more exotic way, but there is no such
need for now.