Make a best effort to delete rules, when disabling monitoring
Since rules may refer to multiple resources, it will not always
be possible to delete rules when disabling monitoring for a single
machine. In that case, we attempt to just remove the given machine
from the list of resources that a rule refers to.

We also no longer delete rules after disassociating metrics from a
machines. Instead, we should perform such action when undeploying
a custom plugin.