Add a unique index on the `key` primary key of PeriodicTaskInfo
First of all, what's seems off here is that the `key` field, even
though it is the documents' primary key, is not indexed. The manual
index on `key`, which was initially set as a work-around, did not
enforce `key` to also be unique, which should be by default the case
for the primary key. This is solved by setting `unique=True` on the
field's definition.

However, setting `unique=True` should not have been required in order
to enforce uniqueness on a primary key. For now, it looks like we've
hit a mongoengine bug when it comes to custom primary keys.