Add an index on the Organization's `name` field
The Organization mongoengine class' clean method performs the following query:

    Organization.objects(name=self.name, id__ne=self.id)

everytime an Organization document is saved. It looks like the above query can
potentially even scan the entire `owner` collection, until a match is found,
which is quite inefficient given a large number of documents and queries executed
on non-indexed fields.

In such cases, query execution times can be up to even 4 seconds! Hopefully, adding
an index on the `name` field will cause the queries to lookup the indexes, instead
of performing a full collection scan.

Ideally, the aforementioned query should be substituted with a unique index on name.
However, this is not possible atm due to Organizations without a name set at all.