Accept either a size id or a dict with size metadata in create_machine
This is useful for providers that do not support pre-defined sizes in
order to avoid having a whole different set of expected parameters.

This way, we will always expect to receive a `size` parameter at the
API level. The `size` may either be an isinstance of basestring, in
which case we trust that it represents the actual size's id, or an
isinstance of dict, in which case it should contain the size's metadata,
like cpu, ram, etc.

For now, this is applicable in only two cases: OnApp and vSphere