Fix threshold check in `rules.plugins.methods:compute`
Since a rule's threshold (as defined in rules.models.main.conditions.QueryCondition.threshold)
is a mongonengine DynamicField, its value can be of any type, eg. string, int, float, etc.
This caused the alert service to compare the datapoints (floats) returned by the TSDB against
the rule's threshold as a string.

In comparisons of numeric and non-numeric objects, the numeric object always comes first:

  - 1 < 'a'
    True
  - 1 < []
    True
  - 0 > {}
    False