Store a resource's creator and owner
Initially, the creator of a resource is also its owner. In terms
of RBAC, a resource's owner has full access on it, even if access
is not explicitly granted by the user's teams' RBAC policies. That
means that as long as a user has permission to create a resource
of any type, he will have Owner rights on it.

For instance, if a team allows all its users to provision VMs, but
"read" access is denied, it means that each user will be able to
"read" the VMs he has provisioned *only*, given he's the creator/owner.