Sharding v1
This commit implements a mixin class to make a scheduler process scalable.

Any subclass of `celerybeatmongo.schedulers.MongoScheduler` can subclass
the `ShardManagerMixin` in order to create a scalable implementation of
the scheduler. The only requirement is that the corresponding scheduler's
`Model` class attribute is set to a class that subclasses the `ShardedScheduleMixin`,
which provides the additional fields necessary for shards' assignment.

For now, each replica of the scheduler attempts to claim as many documents
as possible by assigning them his corresponding shard id.