Fix `me.DynamicField` bug in rules/models/conditions.py
The `value` field should be defined as a DynamicField, since it
should be able to store values of various types, such as string,
integer, or boolean. This is especially applicable in the case
of rules on logs. Since the `filters` EmbeddedDocument has not
been used yet, it has never been persisted to mongoDB, so we can
swap the fields' definitions without the need for a db migration.