Extend the add_rule API endpoint to permit rules on logs
Extension to the request's body include:

  * data_type: the type of the data on which the rule is
    going to operate. For now, acceptable values are "metrics"
    and "logs". For backwards compatibility, if not declared,
    the data_type will be set equal to `metrics`.

  * resource_type: the type of the resource on which the
    rules will be applied, eg. machine, script, etc. For
    backwards compatibility, if no resource_type is set,
    it will be set equal to `machine`. Note that resource_type
    does NOT apply to arbitrary rules and it cannot be updated.

If `selectors` is present in the request body (default), then
a resource-bound rule will be set, otherwise an arbitrary one.
Note that even an empty list of selectors leads to a resource-
bound rule to be created. The empty list of selectors will end
up matching all the resources of the given resource_type.

Based on the above, the API will default to adding a `MachineMetricRule`,
if no other rule type has been explicitly specified.

Rules types are the concatenation of two strings: <str1>-<str2>,
where str1 denotes whether a rule is arbitrary or not and str2
equals the `_data_type_str` class attribute of the respective
Rule subclass. Rules types are just a means to categorize rules.